requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pymysql>=1.1.0
python-dotenv>=1.0.1
PyJWT>=2.8.0
bcrypt>=4.1.0
python-multipart>=0.0.9
